requirements.txt
# Core dependencies
pandas>=1.5.0
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.11.0

# API and networking
kiteconnect>=5.0.0
requests>=2.28.0

# Configuration and data handling
pyyaml>=6.0
python-dotenv>=0.19.0

# Logging and progress
loguru>=0.6.0
tqdm>=4.64.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Optional: Jupyter for research
jupyter>=1.0.0
ipykernel>=6.0.0

# Optional: Additional analysis
scipy>=1.9.0
scikit-learn>=1.1.0 